requirements.txt
Flask==2.1.1
mysql-connector-python
